% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{compute_max_min_ln_marl}
\alias{compute_max_min_ln_marl}
\title{Compute max-min log likelihood value}
\usage{
compute_max_min_ln_marl(u, N)
}
\arguments{
\item{u}{List with useful precomputed information about choice universes}

\item{N}{Matrix containing observed choice counts in a choice experiment.}
}
\value{
Max-min log likelihood value
}
\description{
Compute the max-min log likelihood value, which is both the Random Choice
(RC) log likelihood evaluated at the RC model where all choice distributions
are discrete uniform, and the Random Preference (RP) log likelihood evaluated
at the RP model where all preferences are equally likely. See Section 4.1
of the reference below.
}
\examples{
n <- 5
u <- create_universe(n)
N <- vectorize_counts(u, RanCh::MMS_2019_counts[1, , ])
compute_max_min_ln_marl(u, N)

}
\references{
McCausland, W. (2024). Sequential Monte Carlo for Random Prefernces. Unpublished manuscript.
}
\author{
William McCausland, \email{william.j.mccausland@umontreal.ca}
}
