% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{compute_P_ln_maxl}
\alias{compute_P_ln_maxl}
\title{Compute Random Choice (RC) model P maximising the RC likelihood and the
maximum value of the RC log likelihood.}
\usage{
compute_P_ln_maxl(u, N)
}
\arguments{
\item{u}{List with useful precomputed information about choice universes}

\item{N}{Matrix containing observed choice counts in a choice experiment.}
}
\value{
A list with the following elements
\describe{
\item{ln_maxl}{Maximum value of RC log likelihood}
\item{P_mle}{RC model P maximizing the RC log likelihood}
}
}
\description{
Compute Random Choice (RC) model P maximising the RC likelihood and the
maximum value of the RC log likelihood.
}
\examples{
n <- 5
u <- create_universe(n)
N <- vectorize_counts(u, RanCh::MMS_2019_counts[1, , ])
compute_max_min_ln_marl(u, N)

}
\references{
McCausland, W. (2024). Sequential Monte Carlo for Random Prefernces. Unpublished manuscript.
}
\author{
William McCausland, \email{william.j.mccausland@umontreal.ca}
}
