% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_maxl.R
\name{compute_pi_ln_like}
\alias{compute_pi_ln_like}
\title{Evaluate random preference (RP) log likelihood}
\usage{
compute_pi_ln_like(pi, u, N)
}
\arguments{
\item{pi}{Vector of preference probabilities of length n!}

\item{u}{List with useful precomputed information about choice universes}

\item{N}{Matrix containing observed choice counts in a choice experiment.}
}
\value{
The value of the RP log likelihood at the given value of pi
}
\description{
Evaluate RP log likelihood as a function of pi,
the vector of preference probabilities, for given data N.
}
\examples{
n <- 5
n_fact = factorial(n)
pi <- rep(1/n_fact, n_fact) # Uniform distribution over preference orders
u <- create_universe(n)
N <- vectorize_counts(u, RanCh::MMS_2019_counts[1, , ])
pi_ln_like <- compute_pi_ln_like(pi, u, N)

}
\references{
McCausland, W. (2024). Sequential Monte Carlo for Random Prefernces. Unpublished manuscript.
}
\author{
William McCausland, \email{william.j.mccausland@umontreal.ca}
}
