% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_prior.R
\name{compute_proposal_params}
\alias{compute_proposal_params}
\title{Compute gamma proposal distribution for alpha}
\usage{
compute_proposal_params(u, alpha_prior, N)
}
\arguments{
\item{u}{List with useful precomputed information about choice universes}

\item{alpha_prior}{List created using create_alpha_prior}

\item{N}{Matrix containing observed choice counts in a choice experiment.}
}
\value{
Parameter vector of Beta prime distribution
}
\description{
Find values of the three parameters of the Beta prime distribution that
minimize the Renyi divergence
}
\examples{
library(RanCh)
n <- 5
u <- create_universe(n)
alpha_prior <- create_alpha_prior(n, 4, 0.1)
N <- vectorize_counts(u, RanCh::MMS_2019_counts[1, , ])
theta <- compute_proposal_params(u, alpha_prior, N)

}
