% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{compute_uniform_P_ln_marl}
\alias{compute_uniform_P_ln_marl}
\title{Compute the log marginal likelihood for the uniform-P model}
\usage{
compute_uniform_P_ln_marl(u, N)
}
\arguments{
\item{u}{List with useful precomputed information about choice universes}

\item{N}{Matrix containing observed choice counts in a choice experiment.}
}
\value{
The value of the log marginal likelihood
}
\description{
Compute the log marginal likelihood for the Bayesian Random Choice model
where choice probability vectors are mutually independent and each is
uniformly distributed on its corresponding probability simplex.
}
\examples{
library(RanCh)
n <- 5
u <- create_universe(n)
N <- vectorize_counts(u, RanCh::MMS_2019_counts[1, , ])
compute_uniform_P_ln_marl(u, N)

}
\references{
McCausland, W. (2024). Sequential Monte Carlo for Random Prefernces. Unpublished manuscript.
}
\author{
William McCausland, \email{william.j.mccausland@umontreal.ca}
}
