% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_prior.R
\name{create_alpha_prior}
\alias{create_alpha_prior}
\title{Precompute information for a prior distribution of alpha}
\usage{
create_alpha_prior(n, a, b, h = 0.05, eps = 1e-07)
}
\arguments{
\item{n}{Number of elements in choice universe}

\item{a}{Shape parameter of a Gamma prior distribution of alpha}

\item{b}{Rate parameter of a Gamma prior distribution of alpha}

\item{h}{Distance between grid points in a grid of values of p_bar}

\item{eps}{the probability defining two extreme prior quantiles of alpha}
}
\value{
A list with the following elements
\describe{
\item{a}{Same as input with that name.}
\item{b}{Same as input with that name.}
\item{h}{Same as input with that name.}
\item{p_grid}{Grid of values of p_bar}
\item{alpha_mode}{Grid of values of mode of alpha|pi as function of p_bar}
\item{psi_diff}{Grid of values of psi(1+alpha_mode) - psi(1+alpha_mode/n!)}
\item{p_min}{Minimum value of p_bar in p_grid}
\item{funcs}{Grids of values of alpha, prior cdf and prior pdf}
}
}
\description{
Precompute information assocated with a prior distribution of alpha, including
lookup tables used for Metropolis-Hastings updates of the conditional posterior
distribution of alpha given pi, described in Appendix B of the reference below.
}
\examples{
n <- 5
a <- 4
b <- 0.1
alpha_prior <- create_alpha_prior(n, a, b)

}
\references{
McCausland, W. (2024). Sequential Monte Carlo for Random Prefernces. Unpublished manuscript.
}
\author{
William McCausland, \email{william.j.mccausland@umontreal.ca}
}
