% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{create_universe}
\alias{create_universe}
\title{Precompute tables and matrices depending only on n, the number of objects}
\usage{
create_universe(n)
}
\arguments{
\item{n}{Number of elements in choice universe}
}
\value{
A list with the following elements
\describe{
\item{n}{Same as input with that name.}
\item{n_orders}{Number of preference orders, equal to n!}
\item{n_subsets}{Number of non-empty subsets of choice universe}
\item{n_probs}{Total number of choice probabilities, over all doubleton and larger subsets}
\item{orders}{List of preference orders, with each order a numeric n-vector}
\item{order_strings}{List of strings (e.g. 'abc', 'acb') representing preference orders}
\item{inv_orders}{List of inverse (in the sense of permutation inverse) of preference orders}
\item{singletons}{Vector of singletons, with sets coded as bit strings}
\item{Ax_table}{n_probs by 2 matrix. Each row gives a unique pair of choice set and element}
\item{Ax_strings}{Vector of strings naming rows of Ax_table; the string 'ab;a', for example,
identifies the pair consisting of choice set \{a,b\} and choice object a}
\item{A_table}{n_subsets by 2 matrix. Each row gives the number of choice probabilities associated
with the choice subset and the index of the first row of Ax_table where these choice probabilities
are stored.}
\item{A_strings}{Vector of strings naming rows of A_table; the strings 'a' and
'ab', for example, identify the subsets \{a\} and \{a,b\}}
\item{pi_to_P}{Matrix giving choice probabilities when left multiplying preference probabilities}
\item{pi_to_P_logical}{Matrix, same as pi_to_P but with TRUE/FALSE instead of 1/0}
}
}
\description{
For a choice universe of size n, precompute a list of frequently used
quantities, including tables and convenient row names. The result is a
required input for many functions in this package.
}
\examples{
create_universe(3)      # Example where output is not excessively long
u = create_universe(5)  # Example corresponding to datasets in reference, above.

}
\references{
McCausland, W. (2024). Sequential Monte Carlo for Random Prefernces. Unpublished manuscript.
}
\author{
William McCausland, \email{william.j.mccausland@umontreal.ca}
}
