% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{run_RC_sim}
\alias{run_RC_sim}
\title{Use SMC to simulate the target distribution alpha|N of the Dirichlet RC model}
\usage{
run_RC_sim(u, J, M, alpha_prior, N)
}
\arguments{
\item{u}{List with useful precomputed information about choice universes}

\item{J}{Number of independent groups of particles}

\item{M}{Total number of particles, must be multiple of J}

\item{alpha_prior}{List with information on the prior distribution of alpha}

\item{N}{Matrix containing observed choice counts in a choice experiment.}
}
\value{
A list with the following elements
\describe{
\item{alpha}{A M/J x J matrix with a sample from the target distribution}
\item{marl_stats}{Simulation statistics for the marginal likelihood}
\item{n_plus}{Number of non-zero counts in data}
\item{theta}{Parameter of Beta-Gamma importance distribution}
}
}
\description{
Use SMC to simulate the target distribution alpha|N, the posterior distribution
alpha|N of the scalar parameter alpha of the Dirichlet RC model.
}
\examples{
library(RanCh)
n <- 5
u <- create_universe(n)
alpha_prior <- create_alpha_prior(n, 4, 0.1)
N <- vectorize_counts(u, RanCh::MMS_2019_counts[1, , ])
RC_sim <- run_RC_sim(u, 20, 1000, alpha_prior, N)
}
