% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{run_RP_sim}
\alias{run_RP_sim}
\title{Use SMC to simulate posterior distributions of Dirichlet RC and hybrid models}
\usage{
run_RP_sim(u, J, M, alpha_prior, N, lambda_values, cycle_schedule)
}
\arguments{
\item{u}{List with precomputed information about choice universes
of size n, created using \code{\link[=create_universe]{create_universe()}}}

\item{J}{Number of independent groups of particles}

\item{M}{Total number of particles, must be multiple of J}

\item{alpha_prior}{List with information on the prior distribution of alpha}

\item{N}{Matrix containing observed choice counts in a choice experiment.}

\item{lambda_values}{Vector of indices of the hybrid models}

\item{cycle_schedule}{Schedule of parameters used for the various SMC cycles}
}
\value{
List with the following elements
\describe{
\item{alpha}{A M/J x J matrix with a sample of alpha from the final target distribution}
\item{gamma}{?WJM? sample of gamma from the final target distribution}
\item{lambda_stats}{?WJM? simulation statistics}
\item{cycle_stats}{?WJM? simulation statistics}
\item{big_aPr}{Acceptance probabilities for big blocks}
\item{sm_aPr}{Acceptance probabilities for small blocks}
\item{alpha_aPr}{Acceptance probabilities for alpha proposals}
\item{alpha_mu}{?WJM?}
}
}
\description{
Use SMC to simulate posterior distributions of Dirichlet RC and hybrid models
}
\examples{
library(RanCh)
n <- 5
u <- create_universe(n)
alpha_prior <- create_alpha_prior(n, 4, 0.1)
N <- vectorize_counts(u, RanCh::MMS_2019_counts[1, , ])
J <- 20
M <- 1000
}
