% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC_vec.R
\name{vectorize_counts}
\alias{vectorize_counts}
\title{Vectorize a choice count matrix}
\usage{
vectorize_counts(u, choice_counts)
}
\arguments{
\item{u}{List with useful precomputed information about choice universes}

\item{choice_counts}{Matrix of choice counts indexed by choice set (row) and
choice object (column)}
}
\value{
List with the following elements
\describe{
\item{by_Ax}{Vector of choice counts, indexed by Ax, a flat index coding
both choice set and choice object}
\item{by_A}{Vector of choice counts, indexed by choice set
(aggregates by_Ax)}
}
}
\description{
Create a list with two vectors of choice counts
}
\examples{
n <- 5
u <- create_universe(n)
N <- vectorize_counts(u, RanCh::MMS_2019_counts[1, , ])
}
